refactor(jellyfish-transaction): move vault tx into its dedicated file
